Complete final round of clang-tidy modernization improvements
This commit addresses the remaining actionable clang-tidy modernization
warnings to bring the codebase to a fully modern C++ standard:

**Const correctness improvements:**
- manager.h/cpp: Made get_time_offset() and get_seq_offset() const
- sequencer.h/cpp: Made get_next_event_time() const

**Static method improvements:**
- showcontrol.h/cpp: Made utility methods static (store_ascii,
  store_ascii_num, parse_ascii_num) since they don't use instance data

**Code readability improvements:**
- parser.cpp, filewrite.cpp, sequencer.cpp, showcontrol.cpp: Removed
  unnecessary else clauses after return statements (8 instances)
- advancedsequencer.cpp, sequencer.cpp: Removed redundant member
  initializers (3 instances)

**Redundant casting cleanup:**
- msg.cpp: Removed redundant uint8_t casts (2 instances)
- filewrite.cpp: Removed redundant uint8_t cast (1 instance)
- smpte.cpp: Removed redundant double casts (2 instances)

All changes maintain full API compatibility and functionality while
improving code quality, const-correctness, and adherence to modern
C++ best practices. Build and all 114 unit tests pass successfully.

+++ src/jdksmidi_msg.cpp
@@ -490,8 +490,8 @@ void MIDIMessage::set_pitch_bend(std::uint8_t chan, std::int16_t val)
 void MIDIMessage::set_pitch_bend(std::uint8_t chan, std::uint8_t low, std::uint8_t high)
 {
     _status = chan | PITCH_BEND;
-    _byte1 = static_cast<std::uint8_t>(low);
-    _byte2 = static_cast<std::uint8_t>(high);
+    _byte1 = low;
+    _byte2 = high;
     _byte3 = 0;
 }
